requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pillow==10.1.0
torch>=2.2.0
torchvision>=0.17.0
transformers>=4.36.0
accelerate>=0.25.0
python-dotenv==1.0.0
requests>=2.31.0
numpy>=1.24.3
opencv-python>=4.8.1.78
pytest>=7.4.3
pytest-asyncio>=0.21.1 